# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a VideoWeb API",
        "title": "VideoWeb",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/video/delete": {
            "delete": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video API"
                ],
                "summary": "用户删除视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频ID",
                        "name": "VideoID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/video/download": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Video API"
                ],
                "summary": "用户下载视频(根据视频ID下载视频)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户要下载的视频ID",
                        "name": "VideoID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/video/getVideoDetail": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video API"
                ],
                "summary": "获取对应视频信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "视频ID",
                        "name": "VideoID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/video/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video API"
                ],
                "summary": "用户上传视频",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "视频标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "视频",
                        "name": "uploadVideo",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "视频封面",
                        "name": "videoCover",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "娱乐",
                            "教育",
                            "科技",
                            "知识",
                            "健康",
                            "旅行",
                            "探险",
                            "美食",
                            "时尚",
                            "音乐",
                            "舞蹈",
                            "体育",
                            "健身",
                            "历史",
                            "文化",
                            "游戏",
                            "电影",
                            "搞笑",
                            "资讯"
                        ],
                        "type": "string",
                        "description": "视频分类",
                        "name": "class",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "视频标签",
                        "name": "tags",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "视频描述",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        }
    }
}